Further fixes
Fixed comments on previous commit
