+++ Assets/Scripts/WeatherScripts/ThermometerControl.cs.meta
