--- Assets/WeatherGame/Thermo/ThermometerControl.cs
@@ -1,31 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class ThermometerControl : MonoBehaviour{
-
-    public Text temperatureText;
-    private float updateTemp;
-
-    public Image TempBar;
-    public float maxTemp;
-    public float minTemp;
-
-
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update(){
-        temperatureText.text = (int)updateTemp + " *C;";
-        TempBar.fillAmount = (updateTemp-minTemp) / (maxTemp-minTemp);
-    }
-
-    public void setTemp(float temp){
-        updateTemp = temp;
-    }
-}
